Moving MarkDuplicates to a new file (out of the command-line program).
